+++ PolygonChecker/main.h
@@ -2,4 +2,4 @@
 void printWelcome();
 int printShapeMenu();
 int* getTriangleSides(int* triangleSides);
-bool verifyTriangle(int side1, int side2, int side3);
+bool verifyTriangle(float side1, float side2, float side3);
